luci: upgrade the luci to r23
The "r17" version luci still uses the "atcommand" tool to manage the 3g
modem. It can execute the "AT" command in the luci, but the "atcommand"
is not supported in the pulsar gateway, so the "Signal Quality" always
report the "Unkown" status.

In the "r23" luci, the "comgt" tool is used to get the 3g connection
"Signal Quality" information.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>