multiwan: The 3g-wwan interface can not be found by the multiwan script
The multiwan script mwan_load_config() function could not get the
"network.wwan.ifname" option with the "uci get" tool when the wwan
interface is a "ppp" device such as the "3g" protocol. So add the code
to parse the interface proto, and name it according to the special
device name.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>