dnsmasq: make it still work on a NON UCI system
This bbappend adds a wrapper to parse the dnsmasq configuration
by leveraging the UCI functions. But it will malfunction on a
NON UCI system. So if we include the meta-gateway layer, the
DNS function will not work for all the other cube images except
cube-gw. Adjust this wrapper to fix this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>